feat(cubestore): SQL extension for rolling window queries
Custom syntax is required to capture exact semantics of rolling
window measures from CubeJS. More details on the syntax are in the
commit description inside the DataFusion fork and in the added tests.

Also remove override for `sqlparser-rs` from `Cargo.toml`.
DataFusion specifies the version itself now.

Moved `date_add` to DataFusion in the process.