refactor(@cubejs-client/core): export types and convert type aliases to interfaces
We should be exporting types in general, there's _usually_ no good reason to have them in an `index.d.ts` and then close them off.

Separately, type aliases don't work well with declaration merging, which presents a problem for us as we plan to make use of it. To work around this, all necessary type aliases have been converted to interfaces. These interfaces are then merged with the generated type declarations in `@cubejs-client/dx`.